requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
streamlit
pypdf
wikipedia
python-dotenv
huggingface-hub

